Remove license check from maturity 1
From my point of view the focus on lvl1 should be vulnerabilities and later licenses. I added a "continuous license check" for one of my clients and it is as much work as to include "continuous vulnerability checks". For most of my clients, the _generic impact_ from vulnerabilities is higher than from licenses. For sure, for a software selling company the impact might be higher than for a company which is using libs only for themselves with a "wrong" license.

Please also review my next PR (transitive dependencies) before you response to this suggestion.